backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
langchain==0.3.3
langchain-google-genai==2.0.1
pydantic==2.9.2